pyproject.toml
[tool.poetry]
name = "aiob2"
version = "0.8.0"
description = "A simple and easy to use async wrapper for Backblaze's B2 bucket API."
license = "MIT"
authors = ["Dan <the.void.altacc@gmail.com>"]

readme = "README.md"
repository = "https://github.com/Void-ux/aiob2"
documentation = "https://aiob2.readthedocs.io/en/latest/"
keywords = ["backblaze", "b2", "cloud", "storage"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Void-ux/aiob2/issues/"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.8.1"
typing-extensions = "^4.4.0"

[tool.poetry.extras]
speed = ["orjson"]

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.19.0"
furo = "^2022.12.7"
sphinxcontrib-trio = "^1.1.2"
pytest-order = "^1.0.1"
pytest-env = "^0.8.1"

[tool.pyright]
typeCheckingVersion = "strict"
pythonVersion = "3.8"
reportUnusedImport = "warning"
reportUnnecessaryTypeIgnoreComment = "warning"
reportIncompatibleMethodOverride = "warning"

[tool.pytest.ini_options]
asyncio_mode = "strict"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"